Fix bug in parsing VEP output when aa change goes from 'X' to 'X'
This is likely to be a very rare bug but happens for instance with Mmus:11:60207188:C/- on e!97, which affects
ENSMUST00000134660.7 among others. This is a truncated transcript and the mutation lands on the first annotated
codon (which is truncated) resulting in a NGT to NT change in the codon and being encoded as a X/X amino acid
change. The Ensembl variation API onyl returns the new amino acid when both the REF and ALT are identical.
While in this case the mutation is likely to generate a non-synonymous mutation, both REF and ALT amino acids
are undetermined and are therefore represented with an 'X'.